+++ apps/front/pages/_app.tsx
@@ -1,7 +1,6 @@
 import { AppProps } from 'next/app';
 import Head from 'next/head';
 import './styles.css';
-import { BasicButton } from '../components/forms/Button';
 import { ThemeProvider } from '@mui/material/styles';
 import { theme } from '../style/theme';
 import Header from '../components/nav/header';
@@ -13,7 +12,7 @@ function CustomApp({ Component, pageProps }: AppProps) {
         <title color="primary">Web Chat App</title>
       </Head>
       <main className="app">
-        <Header/>
+        <Header />
         <Component {...pageProps} />
       </main>
     </ThemeProvider>
