+++ apps/front/pages/connexion.tsx
@@ -1,45 +1,44 @@
-import LogForm from "../components/forms/loginform"
-import { useState } from "react";
-import { Box, Container } from "@mui/material";
-import Link from "next/link";
+import LoginForm from '../components/forms/LoginForm';
+import { useState } from 'react';
+import { Box } from '@mui/material';
+import Link from 'next/link';
 
-export default function PageConnexionInscription(){
-    const [formVariant, setFormVariant] = useState<'login' | 'signup'>('login');
+export default function PageConnexionInscription() {
+  const [formVariant, setFormVariant] = useState<'login' | 'signup'>('login');
 
-    const toggleFormVariant = () => {
-        setFormVariant(formVariant === 'login' ? 'signup' : 'login');
-    };
-    return(
-        <Box
-            display="flex"
-            flexDirection="column"
-            alignItems="center" 
-            justifyContent="center"
-        >
-            <LogForm 
-                variant={formVariant} 
-                onLogin={() => {/* Logique de connexion */}} 
-                onSignup={() => {/* Logique d'inscription */}}
-            />
-            <Box mt={2}>
-                {formVariant === 'login' ? (
-                    <>
-                    <Link href="#"  >
-                        Mot de passe oublié ?
-                    </Link>
-                    <Link href="#" onClick={toggleFormVariant}>
-                    Pas encore inscrit ?
-                    </Link>
-                    </>
-                    
-                ) : (
-                    <Link href="#" onClick={toggleFormVariant} >
-                    Déjà inscrit ?
-                    </Link>
-                )}
-            </Box>
-        </Box>
-        
-
-    )
+  const toggleFormVariant = () => {
+    setFormVariant(formVariant === 'login' ? 'signup' : 'login');
+  };
+  return (
+    <Box
+      display="flex"
+      flexDirection="column"
+      alignItems="center"
+      justifyContent="center"
+    >
+      <LoginForm
+        variant={formVariant}
+        onLogin={() => {
+          /* Logique de connexion */
+        }}
+        onSignup={() => {
+          /* Logique d'inscription */
+        }}
+      />
+      <Box mt={2}>
+        {formVariant === 'login' ? (
+          <>
+            <Link href="#">Mot de passe oublié ?</Link>
+            <Link href="#" onClick={toggleFormVariant}>
+              Pas encore inscrit ?
+            </Link>
+          </>
+        ) : (
+          <Link href="#" onClick={toggleFormVariant}>
+            Déjà inscrit ?
+          </Link>
+        )}
+      </Box>
+    </Box>
+  );
 }
