--- apps/front/specs/index.spec.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-
-import Index from '../pages/index';
-
-describe('Index', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<Index />);
-    expect(baseElement).toBeTruthy();
-  });
-});
